refactoring
more refactoring

more clean?

tweaking piechart appearances in golang
